+++ corvis/src/vis/gui.h
@@ -50,8 +50,8 @@ private:
     float scale;
     int width, height;
 
-    std::atomic<int> current_plot{0};
-    std::atomic<int> current_plot_key{-1};
+    std::atomic<size_t> current_plot{0};
+    std::atomic<size_t> current_plot_key{(size_t)-1};
     std::atomic<bool> quit{false};
 
     // Mouse related
